AdminPerms - Move dbDogs tests into its own file
